requirements.txt
asgiref==3.11.1
Django==5.2.11
django-restframework==0.0.1
dotenv==0.9.9
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
